move EventsStream out to its own file
